refactor: 페이지 네이션 글로벌 설정 및 메서드 이름 변경
