+++ src/main/java/com/futurenet/cotree/global/constant/PaginationConstants.java
@@ -1,5 +1,5 @@
-package com.futurenet.cotree.item.policy;
+package com.futurenet.cotree.global.constant;
 
-public class PaginationPolicy {
+public class PaginationConstants {
     public static final int PAGE_SIZE = 20; //한 페이지에 보여줄 상품 개수
 }
