+++ src/main/java/com/futurenet/cotree/item/repository/ItemRepository.java
@@ -10,8 +10,8 @@ import java.util.List;
 @Mapper
 public interface ItemRepository {
     int decreaseStock(@Param("itemId") Long ItemId, @Param("quantity") int quantity);
-    List<Item> findItemsByCategory(@Param("categoryId") Long categoryId, @Param("start") int start, @Param("size") int size);
-    ItemDetail findItemDetailById(@Param("id") Long id);
+    List<Item> getItemsByCategory(@Param("categoryId") Long categoryId, @Param("start") int start, @Param("size") int size);
+    ItemDetail getItemDetailById(@Param("id") Long id);
     List<Item> getEcoItems(@Param("start") int start, @Param("size") int size);
     List<Item> searchItems(@Param("keyword")String keyword, @Param("categoryId") Long categoryId, @Param("start") int start, @Param("size") int size, @Param("isGreen") String isGreen);
 }
